Don't buffer from StringReader
scan(Reader, ...) makes an attempt to ensure that the Reader performs well.
In doing so, it will wrap the Reader with a BufferedReader if it's not already
a BufferedReader, presumably to avoid performance penalties from small reads.

However if it's called with StringReader (or really any number of other Readers)
then there is no upside to buffering, only downside - since BufferedReader will
allocate an 8kb buffer internally, and copy the StringReader's string contents.

Bug: 430962105
Change-Id: If941c0afa6166760c5cfccc1f9797278811f8c55